refactor: move LumberjackFormatter
