+++ libs/ngworker/lumberjack/src/lib/formatting/lumberjack-formatter.service.spec.ts
@@ -8,12 +8,13 @@ import {
   resolveDependency,
 } from '@internal/test-util';
 
-import { LumberjackLogOptions } from './configs';
+import { LumberjackLogOptions } from '../configs';
+import { LumberjackLog } from '../lumberjack-log';
+import { LumberjackLogLevel } from '../lumberjack-log-levels';
+import { LumberjackModule } from '../lumberjack.module';
+import { LumberjackTimeService } from '../time/lumberjack-time.service';
+
 import { LumberjackFormatter } from './lumberjack-formatter.service';
-import { LumberjackLog } from './lumberjack-log';
-import { LumberjackLogLevel } from './lumberjack-log-levels';
-import { LumberjackModule } from './lumberjack.module';
-import { LumberjackTimeService } from './time/lumberjack-time.service';
 
 function createFormatErrorLog(formatterErrorMessage: string, logEntry: LumberjackLog): LumberjackLog {
   return createErrorLog(
